The highscores are responsive, the player's name is asked once
The highscores show on the side when there is space. Otherwise, they are
shown on the top. The player's name is asked only at the start of the game so that the user won't have to keep re-entering their name.

Abnormalies - the highscores might not be responsive on a mobile. When a
player plays the game for the 1st time and their score ranks among the
top 10 high scores, the player might not initially see their score in
the highscores box.